### Changelog category (leave one):
- Performance Improvement


### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Fix contention on Context lock, this significantly improves performance for a lot of short-running concurrent queries.