@kitaisreal 2023-09-29
Benchmark to simulate complex short-running query, this is not worst case scenario.
```
./clickhouse-benchmark --query="SELECT * FROM (SELECT * FROM (SELECT * FROM (SELECT * FROM (SELECT * FROM (SELECT * FROM (SELECT * FROM (SELECT * FROM (SELECT * FROM (SELECT UserID, count(*) FROM (SELECT * FROM datasets.hits_v1 LIMIT 10) GROUP BY UserID))))))))) FORMAT Null" --concurrency=1000
```

Before:

```
localhost:9000, queries: 8698, QPS: 90.463, RPS: 904.634, MiB/s: 0.007, result RPS: 0.000, result MiB/s: 0.000.

0.000%		2.849 sec.
10.000%		9.661 sec.
20.000%		10.214 sec.
30.000%		10.499 sec.
40.000%		10.691 sec.
50.000%		10.851 sec.
60.000%		10.990 sec.
70.000%		11.145 sec.
80.000%		11.306 sec.
90.000%		11.529 sec.
95.000%		11.695 sec.
99.000%		12.049 sec.
99.900%		12.437 sec.
99.990%		12.986 sec.
```
```
SELECT
    quantileExact(0.5)(lock_wait_milliseconds),
    max(lock_wait_milliseconds)
FROM
(
    SELECT (ProfileEvents['ContextLockWaitMicroseconds']) / 1000. AS lock_wait_milliseconds
    FROM system.query_log
    WHERE lock_wait_milliseconds > 0
)

┌─quantileExact(0.5)(lock_wait_milliseconds)─┬─max(lock_wait_milliseconds)─┐
│                                  10428.938 │                   12771.711 │
└────────────────────────────────────────────┴─────────────────────────────┘
```
After:
```
Queries executed: 10177.

localhost:9000, queries: 10177, QPS: 175.670, RPS: 1756.703, MiB/s: 0.013, result RPS: 0.000, result MiB/s: 0.000.

0.000%		0.200 sec.
10.000%		3.185 sec.
20.000%		3.551 sec.
30.000%		4.096 sec.
40.000%		6.402 sec.
50.000%		6.559 sec.
60.000%		6.629 sec.
70.000%		6.697 sec.
80.000%		6.782 sec.
90.000%		6.947 sec.
95.000%		7.848 sec.
99.000%		9.959 sec.
99.900%		12.300 sec.
99.990%		13.092 sec.
```
```
SELECT quantileExact(0.5)(lock_wait_milliseconds), max(lock_wait_milliseconds) FROM (SELECT (ProfileEvents['ContextLockWaitMicroseconds'] / 1000.0) AS lock_wait_milliseconds FROM system.query_log WHERE lock_wait_milliseconds > 0)

┌─quantileExact(0.5)(lock_wait_milliseconds)─┬─max(lock_wait_milliseconds)─┐
│                                      0.001 │                       7.515 │
└────────────────────────────────────────────┴─────────────────────────────┘
```