@kitaisreal 2023-10-08
@rschu1ze it is impossible to check all invariants in Context and we had a lot of bugs even before this refactoring, I started pull request that will use TSA for all Context and Shared context members to find bugs. It is not finished yet, but you can take a look https://github.com/ClickHouse/ClickHouse/pull/55278.